pythonProject27/todo/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.84.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
